modified:   conf/reversetunnel-master-example.json 	new file:   example/reversetunnel/build-tunnel.json 	new file:   example/reversetunnel/build-tunnel.sh 	modified:   src/application/reversetunnel/error.go 	modified:   src/application/reversetunnel/master.go 	modified:   src/application/reversetunnel/masterhttp.go 	modified:   src/application/reversetunnel/protocol.go 	modified:   src/application/reversetunnel/slaver.go

+++ src/application/reversetunnel/slaver.go
@@ -21,7 +21,7 @@ type slaverServer struct {
 type sProxyTunnel struct {
 	rEnd       net.Conn
 	lEnd       net.Conn
-	id         uint64
+	id         uint32
 	targetPort string
 }
